dspy/requirements.txt
dspy>=2.5.0
openai>=1.0.0
anthropic>=0.7.0
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0 